orchestrator-service/pyproject.toml
[project]
name = "orchestrator-service"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "sqlalchemy[asyncio]>=2.0.45",
    "asyncpg>=0.31.0",
    "alembic>=1.18.0",
    "pydantic-settings>=2.12.0",
    "aio-pika>=9.5.8",
    "pyyaml>=6.0.3",
    "openapi-spec-validator>=0.7.2",
]